requirements.txt
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pycodestyle==2.10.0
python-decouple==3.8
requests==2.31.0
tomli==2.0.1
urllib3==2.0.3
